refactor android device manager
